Mined Transactions
This test data has an error rate of 0.3% as it was discovered during testing that two accounts had ether balance that was insufficient for sending transactions. Further the hash.txt file was erroneously has three extra transactions occupying the vacuum left by transactions that geth could not submit by generating a hash for the transaction. However compared to the total number of transactions,10000 the error can be ignored.